Remove `RealtimeChannelInterface`, `ConnectionInterface` and remove empty files
Also refactor class into a separate file
