+++ lib/src/platform/platform.dart
@@ -10,6 +10,7 @@ export 'src/realtime/connection.dart';
 export 'src/realtime/presence.dart';
 export 'src/realtime/realtime.dart';
 export 'src/realtime/realtime_channel.dart';
+export 'src/realtime/realtime_channels.dart';
 export 'src/rest/rest.dart';
 export 'src/rest/rest_channel.dart';
 export 'src/rest/rest_channels.dart';
